gallery changes.

+++ src/views/gallery_page.js
@@ -1,21 +1,31 @@
 import React from "react"
 // import Header2 from "./../header2.js"
-import {Image, CloudinaryContext} from 'cloudinary-react'
+import {Image, CloudinaryContext, Transformation } from 'cloudinary-react'
 import axios from "axios"
-// import base64 from "base-64"
-import FooterPage from "./footer.js"
-import Header from "../header.js"
-import "./gallery.scss"
-
+const cloudinary = require('cloudinary/lib/cloudinary');
 
 export default class GalleryPage extends React.Component {
     constructor() {
         super()
         this.state = {
             gallery: []
-            //will run into problem later of gallery not undefined. need to below do gallery && in the CloudinaryContext? 
         }
+        this.doDelete = false;
+        this.initialValue();
     }
+
+    initialValue(){
+        cloudinary.config({
+            cloud_name: process.env.CLOUD_NAME, 
+            api_key: process.env.CLOUD_API_KEY , 
+            api_secret: process.env.CLOUD_API_SECRET
+        });
+        this.cloudName = process.env.CLOUD_NAME;
+       
+        this.uploadPreset = 'j0ppjpkw';
+        // this.uploadPreset = 'wq6lajqj';
+    }
+
     componentDidMount() {
         // axios.get('https://res.cloudinary.com/dadewebdev/image/fetch/raw_barbershop.json')
         //     .then(res => console.log(res))
@@ -28,7 +38,6 @@ export default class GalleryPage extends React.Component {
         }
         axios({
             url: 'http://localhost:5000/uploadingToGallery/getImages',
-
             method: 'GET',
         }).then(res => {
             this.setState({ gallery: res.data})
@@ -37,17 +46,19 @@ export default class GalleryPage extends React.Component {
 
     uploadWidget() {
         window.cloudinary.openUploadWidget({
-            cloudName: process.env.CLOUD_NAME,
-            uploadPreset: 'wq6lajqj',
-            tags: ['raw_barbershop']
+            cloudName: this.cloudName,
+            uploadPreset: this.uploadPreset,
+            tags: ['raw_barbershop'],
+            api_key: '361324184914465', 
+            api_secret: 'mNefoSj-2o74M74hhvCXXJ3HQAk'
         }, (error, result) => {
+            console.log(result);
             if(result&&result.event=='success'){
                 this.postImages({
                     imageName: result.info.original_filename,
                     imageData: result.info.secure_url,
-                    imagePath: result.info.path,
-                    slug: result.info.path
-                })
+                    imagePublicId: result.info.public_id
+                });
             }
             if (error) console.log(error)
         });
@@ -63,23 +74,24 @@ export default class GalleryPage extends React.Component {
             data: result
         }).then(res => {
             console.log(res);
+            if(res.statusText=="OK"&&this.doDelete){
+                this.deleteImage(res.data.imagePublicId, res.data._id);
+            }
         }).catch(console.log);
     }
 
-    updateImages(result,id){
-        const config = {
-            headers: {'Access-Control-Allow-Origin': '*'}
-        }
-        axios({
-            url: 'http://localhost:5000/uploadingToGallery/updateImage/${id}',
-            method: 'PUT',
-            data: result
-        }).then(res => {
-            console.log(res);
-        }).catch(console.log);
+    updateImages(publicId, id){
+        this.doDelete=true;
+        this.uploadWidget();
     }
 
-    deleteImage(publicId, id){
+
+     deleteImage(publicId, id){
+        console.log(cloudinary);
+        cloudinary.uploader.destroy(publicId
+            ,function(res){
+                console.log(res);
+                if(res.result=="ok"){
                     axios({
                         url: 'http://localhost:5000/uploadingToGallery/deleteImage/'+id,
                         method: 'DELETE',
@@ -89,33 +101,29 @@ export default class GalleryPage extends React.Component {
                 }
             }
         );
-    }	    
+    }
 
     render() {
-
         return (
             <>
-            {/* <Header2/> */}
-            <Header/>
-            <div className="gallery-container" 
-                // style={{backgroundImage: 'url(' + require('./images/haircut_pic.jpg') + ')'}}
-                >
-              
-                    <div className="gallery-content">
-                    <h1>Photo Gallery</h1>
-                    <button onClick={this.uploadWidget.bind(this)}>
-                        Upload Image
-                    </button>
+             {/*<Header2/>*/} 
             <CloudinaryContext cloudName="dadewebdev">
-                {this.state.gallery.map(photo => (
-                <Image publicId={photo.slug} width="200" crop="scale" />
-                // <Transformation width="200" crop="scale" angle="10"/>
+            {this.state.gallery.map(photo => (
+                <Image publicId={photo.slug} />
             ))}
-            </CloudinaryContext>   
-                </div>
+            </CloudinaryContext>      
+                <div className="about-container" 
+                style={{backgroundImage: 'url(' + require('./images/haircut_pic.jpg') + ')'}}
+                >
+                    <div className="content">
+                    <h1>Gallery</h1>
+                    <button
+                        onClick={this.uploadWidget.bind(this)}
+                    >Upload Image</button>
+                    <p>WAITING FOR PICTURES FROM CLIENT</p> 
+                    </div>
 
                 </div>
-                <FooterPage/>
             </> 
         )
     }
